Cherry-pick upstream patches to fix login failure after a failed password entry
debian/patches/0001-libgdm-Don-t-unref-a-connection-that-s-in-use.patch,
debian/patches/0002-libgdm-add-weak-pointer-for-connection-object.patch:
Cherry-pick two patches from the 3.28 branch. These fix a refcounting bug
in the GDM Client object, which had meant that the DBus connection was
being disposed too early in some cases, particularly when logging in after
a failed attempt. The symptom of that was that the system hung on a
"blank" screen. (LP: #1766137)

Fixes LP: #1766137

(bzr r459)