Remove revert_nvidia_wayland_blacklist.patch
That patch was removing the upstream fix for LP: #1705369. And the bug
which the patch was meant to solve doesn't seem to exist any more. If it
did still exist then it should probably be solved in a different way
anyway.

Now Nvidia desktops with multiple GPUs get a working login screen.

Closes: https://bugs.launchpad.net/bugs/1705369