Remove spurious gjs build-dependency
This is less help than we had anticipated in making gdm3 unavailable on
architectures where gnome-shell is unavailable, because gjs mostly works
on s390x but gnome-shell does not. It is also desirable to build libgdm
on even the architectures that do not have a working gdm, so that they
can use gnome-panel and hence gnome-flashback.