debian/patches: Use systemd journal by default
This was already the case with autotools, as the option was set to auto
and enabled when libsystemd was available (which is always the case now).

So import the upstream-proposed patches that remove the request for
never-existed journald pkg-config dependency, and set the
systemd-journal option true by default (using a patch, as it's what also
upstream is meant to do, without bothering debian/rules).