debian/gdm3.gdm-smartcard-*: Ignore invalid user errors on pam_succeed_if.so
We may call the gdm-smartcard module without an user, leaving the module
to figure it out depending on the smartcard certificate.

So we need to ignore PAM_USER_UNKNOWN errors on pam_suceed_if.so.

While pam_sss.so already checks for the user being non root internally,
it's always better to ensure early this in all the cases.

In the pkcs11 case instead we need to check it again after the module
has returned.

LP: #1917362