debian/gdm3.gdm-smartcard-*: Check for /nologin only when an user is defined
pam_nologin.so requires a PAM_USER to be defined in order to check if
the request has been done by root, possibly stopping the login otherwise.
And in case none was provided, it will trigger the fallback pam prompt.

However, with smartcard authentication we may initiate the PAM session
without an user defined and leave to the smartcard service to try to
figure it out depending on the token that has been inserted, that may
have an user associated with it.

So, ensure that we load all the PAM modules that require an user after
the smartcard one, that in case will set one for us.

Only after that, we can fail in case /var/run/nologin is present

LP: #1917362