Add Recommends on gnome-session | x-session-manager
gdm3 already Depends on gnome-session | x-session-manager |
x-window-manager | x-terminal-emulator, but not all packages that
provide x-window-manager are usable as desktop environments (in
particular, the example window manager in the mutter package does not
have a built-in way to launch programs). Adding a Recommends here
provides a stronger hint that nearly all gdm3 users are going to want
an implementation of x-session-manager, preferably GNOME's.

Closes: #986046