Drop default.desktop session
Its name is deeply misleading when it hasn't been GDM's default for
several years, and GDM/GNOME is increasingly moving towards Wayland
rather than X11. It also seems undesirable to present users with an
option that will always need to have a vague, generic name, because we
cannot predict what specific session or window manage it will launch.

Instead, provide an example file with a more appropriate name, which
users of ~/.xsession can either install as-is, or use as a basis for
their own custom session types as documented in gnome-user-docs.