d/generate-config: Change uid to Debian-gdm when acting on /var/lib/gdm3
Writing into a non-root-owned directory as root increases the risk of
symlink or hardlink-based attacks if the non-root user gets compromised.

We historically did a recursive chown on /var/lib/gdm3, so on existing
installations, all files in it (except possibly the dconf database,
which was recreated as root during gdm startup) can be assumed to be
owned by Debian-gdm already. On new installations, all we need to do
is to make sure that if we create /var/lib/gdm3 anew, then we set its
owner correctly. We could do this with systemd-tmpfiles, but since we're
executing a script (d/generate-config) as root during service startup
anyway, the script might as well take responsibility.

`dconf compile` does an atomic overwrite via renaming, so it doesn't
matter if there is already a dconf database owned by root: as long as
Debian-gdm owns the parent directory, the overwrite operation will be
allowed.

In a similar spirit, stop applying a recursive chown to /var/lib/gdm3
in the postinst, and stop shipping root-owned directories in the .deb
that will become owned by Debian-gdm on the running system: after root
has created /var/lib/gdm3 owned by Debian-gdm, all subsequent
manipulation of that directory should be done by Debian-gdm.

In the case of gdm, the Debian-gdm user is already security-sensitive
(because it is responsible for presenting the login prompt, giving it
access to users' passwords), so hardening against it getting compromised
is not a huge practical improvement, but safer is better.