Only build gdm3 on architectures with working mozjs102 and gjs
gjs is likely to be removed from armel as part of the move to mozjs102
(see #1018076), and is unbuildable or fails tests on most non-release
architectures. gdm3 requires gnome-shell at runtime, which requires gjs.
To avoid having an uninstallable gdm3 package, only build it on
architectures where gjs is known to be available.