debian/gdm3-gdm-smartcard*: Do not fail if pam_succeed_if suceeded
We were not handling the success case in pam_succeed_if.so, and so even
if other modules were successful, gdm-smartcard was failing with a
permission denied error, because the pam_succeed_if default was bad, and
this was applied to the success case too.

Alternatively we could even just use success=ignore here, but it's
better to be consistent with other usages.

LP: #1999884