d/gdm3.links: Create a dconf profile named for the Debian-gdm user
In upstream and Ubuntu gdm, the system username used for the greeter
is "gdm", but in Debian, for historical reasons it is "Debian-gdm".
Before version 45, the dconf profile name "gdm" was hard-coded, but
since version 45 it uses a profile name that matches the system username
(to allow gnome-initial-setup to have its own profile).

This resulted in the gdm greeter trying to use the ordinary gnome.session
instead of gnome-login.session, which caused login to fail when
the gnome-session package was not installed, and caused settings in
/etc/gdm3/greeter.dconf-defaults to be ineffective. Other differences
between gnome.session and gnome-login.session, potentially including
security hardening for the greeter, would also not have been effective.

Closes: #1052374, #1051671, #1051993