debian/gdm3.gdm-smartcard-pkcs11-exclusive.pam: Make Smart Card module really exclusive
pam_pkcs11 may stil return PAM_IGNORE in case of smart-card errors if no
card_only is parameter is provided, potentially leading to an unexpected
login.

This should not be an issue anymore in recent versions of this service
files since we always fallback to password auth, but it's still a bug
since we should not really fallback to password if this service file is
selected.

LP: #2017870